requirements.txt
pyautogui
keyboard
pystray
pillow
